services/ai_assistant/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Google Gemini SDK for AI chat (free tier)
google-genai>=1.0.0

# httpx for calling the HTTP API service (inter-service communication)
httpx>=0.27.0
